% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSingleGenes.R
\name{analyzeSingleGenes}
\alias{analyzeSingleGenes}
\title{Analyze Single Genes}
\usage{
analyzeSingleGenes(genesOfInterest, Species = c("hsapiens", "mmusculus"),
  GSEA_Type = c("simple", "complex"), Sample_Type = "normal",
  Tissue = "all", crossCompareMode = FALSE, nperm = 2000,
  whichCompareGroups = c("all", "normal", "cancer"),
  outputPrefix = "CorrelationAnalyzeR_Output", sampler = FALSE,
  runGSEA = TRUE, topPlots = TRUE, returnDataOnly = FALSE,
  pool = NULL)
}
\arguments{
\item{genesOfInterest}{A vector of genes to analyze.}

\item{Species}{Species to obtain gene names for.
Either 'hsapiens' or 'mmusculus'}

\item{GSEA_Type}{Whether GSEA should consider all msigdb annotations,
or just those in the most popular categories. Should be one of either
'simple' or 'complex'.}

\item{Sample_Type}{Type of RNA Seq samples used to create correlation data.
Either "all", "normal", or "cancer". Can be a single value for all genes,
or a vector corresponding to genesOfInterest.}

\item{Tissue}{Which tissue type should gene correlations be derived from?
Default = "all". Can be a single value for all genes,
or a vector corresponding to genesOfInterest.
Run getTissueTypes() to see available tissues.}

\item{crossCompareMode}{Instead of normal single gene analysis, will generate
correlations for single gene across all tissue-disease groups. GSEA will not be run.
Will only consider user input for returnDataOnly, whichCompareGroups,
outputPrefix, Species, and genesOfInterest.}

\item{nperm}{Number of permutations to run in GSEA. Default is 2000}

\item{whichCompareGroups}{For crossCompareMode, select "all", "normal", or "cancer"
to analyze correlations from the corresponding groups.}

\item{outputPrefix}{Prefix for saved files. Should include directory info.}

\item{sampler}{If TRUE, will only return 100,000 random genesets from either
simple or complex TERM2GENEs. Useful for reducing GSEA computational burden.}

\item{runGSEA}{If TRUE will run GSEA using gene correlation values.}

\item{topPlots}{Logical. If TRUE, myGSEA() will build gsea plots for top correlated genesets.}

\item{returnDataOnly}{if TRUE will return only a dataframe of correlation
values and will not generate any folders or files.}

\item{pool}{an object created by pool::dbPool to accessing SQL database.
It will be created if not supplied.}
}
\value{
A named list of correlation values, corGSEA results,
and visualizations for each gene of interest.
}
\description{
Obtains correlations and corGSEA results for each gene of interest
}
\examples{
genesOfInterest <- c("ATM", "SLC7A11")
correlationAnalyzeR::analyzeSingleGenes(genesOfInterest = genesOfInterest,
                              Species = "hsapiens", returnDataOnly = TRUE,
                              GSEA_Type = "simple",
                              Sample_Type = c("normal", "cancer"),
                              Tissue = c("respiratory", "pancreas"))

genesOfInterest <- c("Brca1")
correlationAnalyzeR::analyzeSingleGenes(genesOfInterest = genesOfInterest,
                              Species = "mmusculus",
                              GSEA_Type = "simple", returnDataOnly = TRUE,
                              crossCompareMode = TRUE,
                              whichCompareGroups = "normal")

}
