% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blackListHuman}
\alias{blackListHuman}
\title{A vector of blacklisted tissue-disease categories for human samples}
\format{An object of class \code{character} of length 10.}
\usage{
blackListHuman
}
\description{
A vector of blacklisted tissue-disease categories for human samples
}
\keyword{data}
