% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blackListMouse}
\alias{blackListMouse}
\title{A vector of blacklisted tissue-disease categories for mouse samples}
\format{An object of class \code{character} of length 1.}
\usage{
blackListMouse
}
\description{
A vector of blacklisted tissue-disease categories for mouse samples
}
\keyword{datasets}
