% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadData}
\alias{downloadData}
\title{Download Gene Correlation Data}
\usage{
downloadData(Species, Sample_Type, geneList)
}
\arguments{
\item{Species}{Species to obtain gene names for.
Either 'hsapiens' or 'mmusculus'}

\item{Sample_Type}{Type of RNA Seq samples used to create correlation data.
Either "All", "Normal_Tissues", or "Tumor_Tissues".}

\item{geneList}{Vector of genes for which data will be extracted.}
}
\value{
Downloads files to package data storage.
}
\description{
Downloads gene correlation data when requested from AWS S3 bucket
}
\examples{
downloadData(Species = "hsapiens", Sample_Type = "Normal_Tissues",
    geneList = c("ATM", "BRCA1"))

}
