% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixStrings.R
\name{fixStrings}
\alias{fixStrings}
\title{Fix strings (helper function)}
\usage{
fixStrings(StringVec)
}
\arguments{
\item{StringVec}{A vector of titles (usually GSEA) to clean for visualizations}
}
\description{
Convert vector of GSEA (or other) names to publication-ready titles
}
