% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGeneVsGeneList.R
\name{geneVsGeneListAnalyze}
\alias{geneVsGeneListAnalyze}
\title{Analyze gene vs gene list relationship}
\usage{
geneVsGeneListAnalyze(pairedGenesList, Species = c("hsapiens",
  "mmusculus"), Sample_Type = c("normal", "cancer"), Tissue = "all",
  outputPrefix = "CorrelationAnalyzeR_Output", plotLabels = T,
  sigTest = T, nPerm = 2000, plotMaxMinCorr = T, onlyTop = F,
  topCutoff = 0.5, autoRug = T, plotTitle = T, returnDataOnly = F)
}
\arguments{
\item{pairedGenesList}{A list, named with primary genes of interest with
vectors of secondary genes to test against OR a string containing the
official MSIGDB name for a gene set of interest.}

\item{Species}{Species to obtain gene names for.
Either 'hsapiens' or 'mmusculus'}

\item{Sample_Type}{Type of RNA Seq samples used to create correlation data.
Either "all", "normal", or "cancer". Can be a single value for all genes,
or a vector where each entry corresponds to a gene of interest.}

\item{Tissue}{Which tissue type should gene correlations be derived from?
Default = "all". Can be a single value for all genes, or a vector where each
entry corresponds to a gene of interest.
 Run getTissueTypes() to see available tissue list.}

\item{outputPrefix}{Prefix for saved files. Should include directory info.}

\item{plotLabels}{If TRUE, correlation histograms will contain labeled lines showing
secondary genes and their correlation values.
If list of secondary genes is large, set this to FALSE or onlyTop to TRUE
to avoid cluttering the plot.}

\item{sigTest}{Should the results be compared against random genes?}

\item{nPerm}{Number of bootstrap sampling events to run during sigTest.}

\item{plotMaxMinCorr}{If TRUE, the top correlated and anti-correlated genes
will be plotted alongside the selected secondary genes.}

\item{onlyTop}{For larger secondary gene lists -- This will filter the
number of secondary genes which are plotted to avoid clutter if plotLabels = TRUE.}

\item{topCutoff}{The value used for filtering if 'onlyTop' is 'TRUE'}

\item{autoRug}{If the size of a secondary gene list > 50, plot lines will be replaced
by an auto-generated rug. To disable this behavior, set to FALSE.}

\item{returnDataOnly}{if TRUE will only return a list containing correlations
and significance testing results if applicable.}
}
\value{
A list containing correlation values and signficance testing results
}
\description{
Explores how a list of secondary genes relates to a primary gene of interest
}
\examples{
pairedGenesList <- list("TP53" = c("BRCA1", "CDK12", "PARP1"),
                        "SON" = c("AURKB", "SFPQ", "DHX9"))

pairedGenesList <- list("ATM" = "PUJANA_BRCA1_PCC_NETWORK")

Result <- pairedGenesAnalyzeR(pairedGenesList = pairedGenesList,
                              Species = "hsapiens",
                              Sample_Type = "Normal_Tissues")

}
