% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelationData.R
\name{getCorrelationData}
\alias{getCorrelationData}
\title{Get Gene Correlation Data}
\usage{
getCorrelationData(Species, Sample_Type, Tissue, geneList)
}
\arguments{
\item{Species}{Species to obtain gene names for.
Either 'hsapiens' or 'mmusculus'}

\item{Sample_Type}{Type of RNA Seq samples used to create correlation data.
Either "All", "Normal_Tissues", or "Tumor_Tissues".}

\item{Tissue}{Type of tissue to extract correlations from. e.g. brain, lung, all.}

\item{geneList}{Vector of genes for which data will be extracted.}
}
\value{
A correlation data frame object
}
\description{
Queries from MySQL the requested gene correlation data
}
\examples{
correlationDF <- getCorrelationData(Species = "hsapiens", Sample_Type = "Normal_Tissues",
                                    geneList = c("ATM", "BRCA1"))

}
