% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTissueTPM.R
\name{getTissueTPM}
\alias{getTissueTPM}
\title{Get TPM values for tissues and gene of interest}
\usage{
getTissueTPM(genesOfInterest, Species = c("hsapiens", "mmusculus"),
  Tissues = "all", Sample_Type = c("all", "normal", "cancer"),
  useBlackList = TRUE, pool = NULL)
}
\arguments{
\item{genesOfInterest}{A length-two vector with genes to compare.}

\item{Species}{Species to obtain tissue types for.
Either 'hsapiens' or 'mmusculus'}

\item{Tissues}{Which tissue type should TPM be collected for? See available options
with getTissueTypes().}

\item{Sample_Type}{Type of RNA Seq samples to obtain TPM for? See available options
with getTissueTypes().}

\item{useBlackList}{Should black-listed tissue/disease categories for this species
be removed from the returned list? Improves the quality of analysis by removing
categories with low sample numbers and high observed variance.}

\item{pool}{an object created by pool::dbPool to accessing SQL database.
It will be created if not supplied.}
}
\value{
List of TPM matrices for each selected tissue-disease combination.
}
\description{
Downloads TPM values for tissues of interest
}
\examples{
res <- getTissueTPM(genesOfInterest = c("BRCA1", "ATM"),
                    Species = "hsapiens",
                    Tissues = c("brain", "respiratory"),
                    Sample_Type = "all",
                    useBlackList = TRUE)
}
