% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTissueTypes.R
\name{getTissueTypes}
\alias{getTissueTypes}
\title{Get tissue types}
\usage{
getTissueTypes(Species = c("hsapiens", "mmusculus"))
}
\arguments{
\item{Species}{Species to obtain tissue types for.
Either 'hsapiens' or 'mmusculus'}
}
\value{
Vector containing available tissue types.
}
\description{
Finds tissue types with available correlation data
}
\examples{
getTissueTypes("hsapiens")

}
