% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTissueVST.R
\name{getTissueVST}
\alias{getTissueVST}
\title{Get VST values for tissues and gene of interest}
\usage{
getTissueVST(
  genesOfInterest,
  Tissues = "all",
  Sample_Type = c("normal", "cancer"),
  useBlackList = TRUE,
  pool = NULL
)
}
\arguments{
\item{genesOfInterest}{A length-two vector with genes to compare.}

\item{Tissues}{Which tissue type should VST be collected for? See available options
with getTissueTypes().}

\item{Sample_Type}{Type of RNA Seq samples to obtain VST for? See available options
with getTissueTypes().}

\item{useBlackList}{Should black-listed tissue/disease categories for this species
be removed from the returned list? Improves the quality of analysis by removing
categories with low sample numbers and high observed variance.}

\item{pool}{an object created by pool::dbPool to accessing SQL database.
It will be created if not supplied.}
}
\value{
List of VST matrices for each selected tissue-disease combination.
}
\description{
Downloads VST values for tissues of interest
}
\examples{
res <- getTissueVST(genesOfInterest = c("BRCA1", "ATM"),
                    Tissues = c("brain", "respiratory"),
                    Sample_Type = "all",
                    useBlackList = TRUE)
}
