% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsapiens_complex_TERM2GENE}
\alias{hsapiens_complex_TERM2GENE}
\title{A tbl that maps gene names to gene sets for all human genesets}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2448016 rows and 2 columns.}
\source{
msigdbr()
}
\usage{
hsapiens_complex_TERM2GENE
}
\description{
A tbl that maps gene names to gene sets for all human genesets
}
\keyword{datasets}
