% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsapiens_corrSmall_geneNames}
\alias{hsapiens_corrSmall_geneNames}
\title{A vector of valid human genes to extract correlations from}
\format{An object of class \code{character} of length 28685.}
\usage{
hsapiens_corrSmall_geneNames
}
\description{
A vector of valid human genes to extract correlations from
}
\keyword{data}
