% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsapiens_simple_TERM2GENE}
\alias{hsapiens_simple_TERM2GENE}
\title{A tbl that maps gene names to gene sets for human genesets from popular MSIGDB categories:
H, C2, C5, C6}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 993906 rows and 2 columns.}
\source{
msigdbr()
}
\usage{
hsapiens_simple_TERM2GENE
}
\description{
A tbl that maps gene names to gene sets for human genesets from popular MSIGDB categories:
H, C2, C5, C6
}
\keyword{datasets}
