% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_coldata}
\alias{human_coldata}
\title{A data frame of sample info from GEO (human)}
\format{
An object of class \code{data.frame} with 81716 rows and 6 columns.
}
\usage{
human_coldata
}
\description{
A data frame of sample info from GEO (human)
}
\keyword{data}
