% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_grouplist}
\alias{human_grouplist}
\title{A list of sample categorizations (human)}
\format{
An object of class \code{list} of length 25.
}
\usage{
human_grouplist
}
\description{
A list of sample categorizations (human)
}
\keyword{data}
