% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanGenes}
\alias{humanGenes}
\title{A data frame that maps human genes to their short-form descriptions}
\format{An object of class \code{data.frame} with 58951 rows and 2 columns.}
\source{
biomaRt
}
\usage{
humanGenes
}
\description{
A data frame that maps human genes to their short-form descriptions
}
\keyword{datasets}
