% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanGenesTPM}
\alias{humanGenesTPM}
\title{A vector of valid human genes with TPM data available}
\format{An object of class \code{character} of length 27994.}
\usage{
humanGenesTPM
}
\description{
A vector of valid human genes with TPM data available
}
\keyword{datasets}
