% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanGenesVST}
\alias{humanGenesVST}
\title{A vector of valid human genes with VST data available}
\format{
An object of class \code{character} of length 28722.
}
\usage{
humanGenesVST
}
\description{
A vector of valid human genes with VST data available
}
\keyword{data}
