% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmusculus_complex_TERM2GENE}
\alias{mmusculus_complex_TERM2GENE}
\title{A tbl that maps gene names to gene sets for all mouse genesets}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2394090 rows and 2 columns.}
\source{
msigdbr()
}
\usage{
mmusculus_complex_TERM2GENE
}
\description{
A tbl that maps gene names to gene sets for all mouse genesets
}
\keyword{datasets}
