% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmusculus_corrSmall_geneNames}
\alias{mmusculus_corrSmall_geneNames}
\title{A vector of valid mouse genes to extract correlations from}
\format{An object of class \code{character} of length 24924.}
\usage{
mmusculus_corrSmall_geneNames
}
\description{
A vector of valid mouse genes to extract correlations from
}
\keyword{data}
