% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouseGenes}
\alias{mouseGenes}
\title{A data frame that maps mouse genes to their short-form descriptions}
\format{An object of class \code{data.frame} with 55619 rows and 2 columns.}
\source{
biomaRt
}
\usage{
mouseGenes
}
\description{
A data frame that maps mouse genes to their short-form descriptions
}
\keyword{datasets}
