% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouseGenesTPM}
\alias{mouseGenesTPM}
\title{A vector of valid mouse genes with TPM data available}
\format{An object of class \code{character} of length 24499.}
\usage{
mouseGenesTPM
}
\description{
A vector of valid mouse genes with TPM data available
}
\keyword{data}
