% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSIGDB_Geneset_Names}
\alias{MSIGDB_Geneset_Names}
\title{A vector of valid MSIGDB geneset names}
\format{An object of class \code{character} of length 17810.}
\source{
msigdbr()
}
\usage{
MSIGDB_Geneset_Names
}
\description{
A vector of valid MSIGDB geneset names
}
\keyword{datasets}
