% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSIGDB_Geneset_Small_Names}
\alias{MSIGDB_Geneset_Small_Names}
\title{A vector of valid MSIGDB geneset names with fewer then 500 genes associated with them}
\format{An object of class \code{character} of length 21940.}
\source{
msigdbr()
}
\usage{
MSIGDB_Geneset_Small_Names
}
\description{
A vector of valid MSIGDB geneset names with fewer then 500 genes associated with them
}
\keyword{data}
