% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGSEA.R
\name{myGSEA}
\alias{myGSEA}
\title{Wrapper for clusterProlifer's GSEA()}
\usage{
myGSEA(ranks, TERM2GENE, padjustedCutoff = 0.05,
  returnDataOnly = FALSE, nperm = 2000, topPlots = TRUE, outDir,
  Condition = "GSEA Results", plotFile = "GSEA_results")
}
\arguments{
\item{ranks}{Numeric of gene 'scores' ordered by decreasing value and
named with gene symbols.}

\item{TERM2GENE}{Data frame with two columns: gene set identifiers and
gene symbols. Can be generated using correlationAnalyzeR::getTERM2GENE()}

\item{padjustedCutoff}{Value to use as a cutoff for returned gene sets.}

\item{returnDataOnly}{Should GSEA data/plots be saved to file? Default: FALSE}

\item{nperm}{Number of permutations to run. Default is 2000}

\item{topPlots}{Should top GSEA pathways be plotted? Default: TRUE}

\item{outDir}{output directory.}

\item{Condition}{Name to use for titles of plots. Default = "GSEA Results".}

\item{plotFile}{prefix to use for naming output files.}
}
\value{
Named list containing GSEA() output, GSEA data frame, and visualizations.
}
\description{
Runs GSEA() from clusterProfiler and creates useful
visualizations.
}
\examples{
corrDF <- correlationAnalyzeR::analyzeSingleGenes(genesOfInterest = c("BRCA1"),
 returnDataOnly = TRUE, runGSEA = FALSE, Sample_Type = "normal")
ranks <- corrDF$correlations[,1]
names(ranks) <- rownames(corrDF$correlations)
TERM2GENE <- correlationAnalyzeR::getTERM2GENE(GSEA_Type = "simple",
Species = "hsapiens")
correlationAnalyzeR::myGSEA(ranks = ranks,
TERM2GENE = TERM2GENE,
       plotFile = "GSEA_out", outDir = getwd(),
       topPlots = FALSE, returnDataOnly=TRUE, Condition = "GSEA Results")

}
