% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGenePairs.R
\name{pairedGenesAnalyzeR}
\alias{pairedGenesAnalyzeR}
\title{Analyze Gene Pairs}
\usage{
pairedGenesAnalyzeR(pairedGenesList, Species = c("hsapiens",
  "mmusculus"), Sample_Type = c("All", "Normal_Tissues",
  "Tumor_Tissues"), outputPrefix = "CorrelationAnalyzeR_Output",
  sigTest = T, plotMaxMinCorr = T, onlyTop = F, topCutoff = 0.5)
}
\arguments{
\item{pairedGenesList}{A list, named with primary genes of interest with
vectors of secondary genes to test against.}

\item{Species}{Species to obtain gene names for.
Either 'hsapiens' or 'mmusculus'}

\item{Sample_Type}{Type of RNA Seq samples used to create correlation data.
Either "All", "Normal_Tissues", or "Tumor_Tissues".}

\item{outputPrefix}{Prefix for saved files. Should include directory info.}

\item{sigTest}{Should the results be compared against random genes?}

\item{plotMaxMinCorr}{If TRUE, the top correlated and anti-correlated genes
will be plotted alongside the selected secondary genes.}

\item{onlyTop}{For larger secondary gene lists -- This will filter the
number of secondary genes which are plotted to avoid clutter.}

\item{topCutoff}{The value used for filtering if 'onlyTop' is 'TRUE'}
}
\value{
A list containing correlation values and signficance testing results
}
\description{
Explores how a list of secondary genes relates to a primary gene of interest
}
\examples{
pairedGenesList <- list("TP53" = c("BRCA1", "CDK12", "PARP1"),
                        "SON" = c("AURKB", "SFPQ", "DHX9"))
Result <- pairedGenesAnalyzeR(pairedGenesList = pairedGenesList,
                              Species = "hsapiens",
                              Sample_Type = "Normal_Tissues")

}
