% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathwayCategories}
\alias{pathwayCategories}
\title{A vector containing the names of valid TERM2GENE categories for GSEA_Type or pathwayType input.}
\format{An object of class \code{character} of length 19.}
\usage{
pathwayCategories
}
\description{
A vector containing the names of valid TERM2GENE categories for GSEA_Type or pathwayType input.
}
\keyword{data}
