% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleTPMOrderHuman}
\alias{sampleTPMOrderHuman}
\title{A vector containing the order of human samples in the sample-tissue SQL table}
\format{An object of class \code{array} of length 74771.}
\usage{
sampleTPMOrderHuman
}
\description{
A vector containing the order of human samples in the sample-tissue SQL table
}
\keyword{datasets}
