% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleTPMOrderMouse}
\alias{sampleTPMOrderMouse}
\title{A vector containing the order of mouse samples in the sample-tissue SQL table}
\format{An object of class \code{array} of length 68683.}
\usage{
sampleTPMOrderMouse
}
\description{
A vector containing the order of mouse samples in the sample-tissue SQL table
}
\keyword{datasets}
