% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleVSTOrderHuman}
\alias{sampleVSTOrderHuman}
\title{A vector containing the order of human samples in the sample-tissue SQL table}
\format{
An object of class \code{character} of length 81716.
}
\usage{
sampleVSTOrderHuman
}
\description{
A vector containing the order of human samples in the sample-tissue SQL table
}
\keyword{data}
