% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGenePairs.R
\name{analyzeGenePairs}
\alias{analyzeGenePairs}
\title{Analyze Gene Pairs}
\usage{
analyzeGenePairs(genesOfInterest, Sample_Type = c("normal", "normal"),
  Tissue = c("all", "all"), Species = c("hsapiens", "mmusculus"),
  GSEA_Type = c("simple", "complex"),
  outputPrefix = "CorrelationAnalyzeR_Output_Paired", runGSEA = T,
  topPlots = F, returnDataOnly = F)
}
\arguments{
\item{genesOfInterest}{A vector with two genes to compare.}

\item{Sample_Type}{A vector of length-2 corresponding to the genesOfInterest.
Indicates the type of sample for each; "all", "normal", or "cancer".
Default: c("normal", "normal")}

\item{Tissue}{A vector of length-2 corresponding to the genesOfInterest.
Indicates the type of tissue for each; Run getTissueTypes() to see available list.
Default: c("all", "all")}

\item{Species}{Species to obtain gene names for.
Either 'hsapiens' or 'mmusculus'. Default: "hsapiens".}

\item{GSEA_Type}{Whether GSEA should consider all msigdb annotations,
or just those in the most popular categories. Should be one of either
'simple' or 'complex'.}

\item{outputPrefix}{Prefix for saved files. Should include directory info.}

\item{runGSEA}{If TRUE will run GSEA using gene correlation values.}

\item{returnDataOnly}{if TRUE will return only a dataframe of correlation
values and will not generate any folders or files.}
}
\value{
A dataframe of correlation values for each gene of interest.
}
\description{
Comaprison of two genes of interest using correlation values.
This can be 2 different genes in the same tissue or sample type. Or
the same gene accross two sample or tissue types.
}
\examples{
genesOfInterest <- c("ATM", "SLC7A11")
Result <- analyzeSingleGenes(genesOfInterest = genesOfInterest,
                              Species = "hsapiens",
                              GSEA_Type = "simple",
                              Sample_Type = "Normal_Tissues")

}
