% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGenePairs.R
\name{analyzeGenePairs}
\alias{analyzeGenePairs}
\title{Analyze Gene Pairs}
\usage{
analyzeGenePairs(genesOfInterest, Sample_Type = c("normal", "normal"),
  Tissue = c("all", "all"), Species = c("hsapiens", "mmusculus"),
  GSEA_Type = c("simple"),
  outputPrefix = "CorrelationAnalyzeR_Output_Paired",
  crossCompareMode = FALSE, runGSEA = TRUE, TERM2GENE = NULL,
  nperm = 2000, sampler = FALSE, topPlots = FALSE,
  returnDataOnly = TRUE, pool = NULL)
}
\arguments{
\item{genesOfInterest}{A length-two vector with genes to compare.}

\item{Sample_Type}{A length-two vector of sample types corresponding to genesOfInterest.
Choose "all", "normal", or "cancer". Default: c("normal", "normal")}

\item{Tissue}{A length-two vector of tissue types corresponding to genesOfInterest.
Run getTissueTypes() to see available list. Default: c("all", "all")}

\item{Species}{Species to obtain gene names for.
Either 'hsapiens' or 'mmusculus'. Default: "hsapiens".}

\item{GSEA_Type}{Which GSEA annotations should be considered? Options listed in
correlationAnalyzeR::pathwayCategories -- See details of getTERM2GENE for more info.}

\item{outputPrefix}{Prefix for saved files -- the directory name to store output files in. If
folder does not exist, it will be created.}

\item{crossCompareMode}{Use this mode to generate comparisons
across all tissue and disease types. If both genes for genesOfInterest are the
same -- will compare normal vs cancer for that gene in each available tissue. Else, will
perform comparison of two different genes in all tissue-disease groups.
Will only consider user input for returnDataOnly, outputPrefix, Species, and genesOfInterest.}

\item{runGSEA}{If TRUE will run GSEA using gene correlation values.}

\item{TERM2GENE}{Mapping of geneset IDs to gene names. If NULL, it will be
generated automatically. Only applicable if GSEA is to be run.}

\item{nperm}{Number of permutations to run in GSEA. Default is 2000}

\item{sampler}{If TRUE, will only return 100,000 random genesets. Useful for reducing GSEA computational burden.}

\item{topPlots}{Logical. If TRUE, myGSEA() will build gsea plots for top correlated genesets.}

\item{returnDataOnly}{if TRUE will return result list object
and will not generate any folders or files.}

\item{pool}{an object created by pool::dbPool to accessing SQL database.
It will be created if not supplied.}
}
\value{
A named list containing visualizations and correlation data from paired analysis.
}
\description{
Comaprison of two genes of interest using correlation values.
This can be 2 different genes in the same tissue or sample type or
the same gene accross two sample or tissue types. Alternatively, specify 'crossCompareMode'
to view compared correrlations across all available tissue types.
}
\examples{
genesOfInterest <- c("ATM", "SLC7A11")
correlationAnalyzeR::analyzeGenePairs(genesOfInterest = genesOfInterest,
                              Species = "hsapiens",
                              GSEA_Type = "simple", returnDataOnly = TRUE,
                              Sample_Type = c("normal", "normal"),
                              Tissue = c("brain", "brain"))
genesOfInterest <- c("BRCA1", "BRCA1")
correlationAnalyzeR::analyzeGenePairs(genesOfInterest = genesOfInterest,
                              Species = "hsapiens",
                              GSEA_Type = "simple", returnDataOnly = TRUE,
                              Sample_Type = c("normal", "cancer"),
                              Tissue = c("respiratory", "respiratory"))
genesOfInterest <- c("NFKB1", "SOX10")
correlationAnalyzeR::analyzeGenePairs(genesOfInterest = genesOfInterest,
                              Species = "hsapiens", returnDataOnly = TRUE,
                              crossCompareMode = TRUE)
}
