% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGeneListTopology.R
\name{analyzeGenesetTopology}
\alias{analyzeGenesetTopology}
\title{Analyze Gene List Topology}
\usage{
analyzeGenesetTopology(genesOfInterest, Species = c("hsapiens",
  "mmusculus"), Sample_Type = c("Normal_Tissues", "Tumor_Tissues"),
  crossComparisonType = c("variantGenes", "coCorrelativeGenes", "PCA"),
  setComparisonCutoff = "Auto", numTopGenesToPlot = "Auto",
  alternativeTSNE = T, numClusters = "Auto",
  outputPrefix = "CorrelationAnalyzeR_Output", returnDataOnly = F)
}
\arguments{
\item{genesOfInterest}{A vector of genes to analyze or the name of an official MSIGDB term.}

\item{Species}{Species to obtain gene names for.
Either 'hsapiens' or 'mmusculus'}

\item{Sample_Type}{Type of RNA Seq samples used to create correlation data.
Either "All", "Normal_Tissues", or "Tumor_Tissues".}

\item{crossComparisonType}{The type of topology tests to run. (see details)}

\item{setComparisonCutoff}{Only relevant for cocorrelation analysis -- the number of genes which
must aggree for a gene to be considered co-correlative within the input gene list.}

\item{numTopGenesToPlot}{When creating a heatmap of the top co-correlative or top variant genes,
how many genes should be plotted on the y axis?}

\item{alternativeTSNE}{Logical. If TRUE, then a TSNE will be run as an alternative to PCA for visualizing
large input gene lists. This is highly recommended as 100+ member gene lists cannot be visualized otherwise.}

\item{numClusters}{The number of clusters to create with hclust or TSNE analysis.}

\item{outputPrefix}{Prefix for saved files. Should include directory info.}

\item{returnDataOnly}{if TRUE will return only a list of analysis results.}
}
\value{
A list of analysis results and plotting data.
}
\description{
Analyzes the topology of a gene list using gene correlation data and dimension-reduction techniques.
}
\details{
Cross Comparison Types:
- variantGenes: These are the genes which best explain variation between genes within the input list.
                These genes can divide a list into functional groups.
- coCorrelativeGenes: These are the genes which best explain similarities between all genes in the input list.
                      These genes can explain what biological processes unify the input genes.
- PCA: This is a dimensionality reduction technique for exploring the topology of a gene list.
       The PCA analyses here employes hclust to divide the gene list into functional clusters.
       If the input list is > 100 genes, RTsne will be used for visualization.
}
\examples{
genesOfInterest <- c("CDK12", "AURKB", "SFPQ", "PARP1", "BRCC3", "BRCA2", "PARP1",
                     "DHX9", "SON", "AURKA", "SETX", "BRCA1", "ATMIN")
Result <- analyzeGenesetTopology(genesOfInterest = genesOfInterest,
                                 Species = "hsapiens",
                                 Sample_Type = "Normal_Tissues",
                                 crossComparisonType = c("variantGenes", "PCA"))

}
