% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSingleGenes.R
\name{analyzeSingleGenes}
\alias{analyzeSingleGenes}
\title{Analyze Single Genes}
\usage{
analyzeSingleGenes(genesOfInterest, Species = c("hsapiens", "mmusculus"),
  GSEA_Type = c("simple", "complex"), Sample_Type = c("Normal_Tissues",
  "Tumor_Tissues"), outputPrefix = "CorrelationAnalyzeR_Output",
  runGSEA = T, returnDataOnly = F)
}
\arguments{
\item{genesOfInterest}{A vector of genes to analyze.}

\item{Species}{Species to obtain gene names for.
Either 'hsapiens' or 'mmusculus'}

\item{GSEA_Type}{Whether GSEA should consider all msigdb annotations,
or just those in the most popular categories. Should be one of either
'simple' or 'complex'.}

\item{Sample_Type}{Type of RNA Seq samples used to create correlation data.
Either "All", "Normal_Tissues", or "Tumor_Tissues".}

\item{outputPrefix}{Prefix for saved files. Should include directory info.}

\item{runGSEA}{If TRUE will run GSEA using gene correlation values.}

\item{returnDataOnly}{if TRUE will return only a dataframe of correlation
values and will not generate any folders or files.}
}
\value{
A dataframe of correlation values for each gene of interest.
}
\description{
Obtains correlation info and GSEA results for each gene of interest
}
\examples{
genesOfInterest <- c("ATM", "SLC7A11")
Result <- analyzeSingleGenes(genesOfInterest = genesOfInterest,
                              Species = "hsapiens",
                              GSEA_Type = "simple",
                              Sample_Type = "Normal_Tissues")

}
