% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGeneVsGeneList.R
\name{geneVsGeneListAnalyze}
\alias{geneVsGeneListAnalyze}
\title{Analyze gene vs gene list relationship}
\usage{
geneVsGeneListAnalyze(pairedGenesList, Species = c("hsapiens",
  "mmusculus"), Sample_Type = c("normal", "cancer"), Tissue = "all",
  outputPrefix = "CorrelationAnalyzeR_Output", plotLabels = TRUE,
  sigTest = TRUE, nPerm = 2000, plotMaxMinCorr = TRUE,
  onlyTop = FALSE, topCutoff = 0.5, autoRug = TRUE,
  plotTitle = TRUE, returnDataOnly = TRUE, pool = NULL,
  makePool = FALSE)
}
\arguments{
\item{pairedGenesList}{A list, named with primary genes of interest with
vectors of secondary genes to test against OR a string containing the
official MSIGDB name for a gene set of interest.}

\item{Species}{Species to obtain gene names for.
Either 'hsapiens' or 'mmusculus'}

\item{Sample_Type}{Type of RNA Seq samples used to create correlation data.
Either "all", "normal", or "cancer". Can be a single value for all genes,
or a vector corresponding to genesOfInterest.}

\item{Tissue}{Which tissue type should gene correlations be derived from?
Default = "all". Can be a single value for all genes,
or a vector corresponding to genesOfInterest.
Run getTissueTypes() to see available tissues.}

\item{outputPrefix}{Prefix for saved files. Should include directory info.}

\item{plotLabels}{If TRUE, correlation histograms will contain labeled lines showing
secondary genes and their correlation values.
If list of secondary genes is large, set this to FALSE or onlyTop to TRUE
to avoid cluttering the plot. Default: TRUE.}

\item{sigTest}{Should the results be compared against random? Default: TRUE.}

\item{nPerm}{Number of bootstrap sampling events to run during sigTest. Default: 2000.}

\item{plotMaxMinCorr}{If TRUE, the top correlated and anti-correlated genes
will be plotted alongside the selected secondary genes. Default: TRUE.}

\item{onlyTop}{For larger secondary gene lists -- This will filter the
number of secondary genes which are plotted to avoid clutter if plotLabels = TRUE.
Default: FALSE.}

\item{topCutoff}{The value used for filtering if 'onlyTop' is 'TRUE'. Default: .5}

\item{autoRug}{If the size of a secondary gene list > 50, plot lines will be replaced
by an auto-generated rug. Default: TRUE.}

\item{plotTitle}{Logical. If TRUE, plot title will be added to visualizations. Default: TRUE.}

\item{returnDataOnly}{if TRUE will only return a list containing correlations
and significance testing results if applicable. Default: TRUE.}

\item{pool}{an object created by pool::dbPool to accessing SQL database.
It will be created if not supplied.}

\item{makePool}{Logical. Should a pool be created if one is not supplied? Default: FALSE.}
}
\value{
A list containing correlation values and signficance testing results
}
\description{
Explores how a list of secondary genes relates to a primary gene of interest
}
\examples{
pairedGenesList <- list("TP53" = c("BRCA1", "CDK12", "PARP1"),
                        "SON" = c("AURKB", "SFPQ", "DHX9"))

correlationAnalyzeR::geneVsGeneListAnalyze(pairedGenesList = pairedGenesList,
                              Species = "hsapiens", returnDataOnly = TRUE,
                              Sample_Type = "normal",
                              Tissue = "brain")

}
