% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAvailableGenes.R
\name{getAvailableGenes}
\alias{getAvailableGenes}
\title{Get Available Genes}
\usage{
getAvailableGenes(pool = NULL)
}
\arguments{
\item{pool}{an object created by pool::dbPool to accessing SQL database.
It will be created if not supplied.}
}
\value{
A vector of genes with associated correlation data
}
\description{
Finds available genes within correlation data
}
\examples{
correlationAnalyzeR::getAvailableGenes()

}
