% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAvailableGenes.R
\name{getAvailableGenes}
\alias{getAvailableGenes}
\title{Get Available Genes}
\usage{
getAvailableGenes(Species = c("hsapiens", "mmusculus"))
}
\arguments{
\item{Species}{Species to obtain gene names for.
Either 'hsapiens' or 'mmusculus'}
}
\value{
Dataframe containing available genes and NCBI descriptions.
}
\description{
Finds available genes within correlation data
}
\examples{
correlationAnalyzeR::getAvailableGenes("hsapiens")

}
