% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelationData.R
\name{getCorrelationData}
\alias{getCorrelationData}
\title{Get Gene Correlation Data}
\usage{
getCorrelationData(Species, Sample_Type, Tissue, geneList, pool = NULL)
}
\arguments{
\item{Species}{Species to obtain gene names for.
Either 'hsapiens' or 'mmusculus'}

\item{Sample_Type}{Type of RNA Seq samples used to create correlation data.
Either "all", "normal", or "cancer". Can be a single value for all genes,
or a vector corresponding to geneList.}

\item{Tissue}{Which tissue type should gene correlations be derived from?
Default = "all". Can be a single value for all genes,
or a vector corresponding to geneList.
Run getTissueTypes() to see available tissue list.}

\item{geneList}{Vector of genes for which data will be extracted.}

\item{pool}{an object created by pool::dbPool to accessing SQL database.
It will be created if not supplied.}
}
\value{
A correlation data frame object
}
\description{
Obtain correlation data by querying MySQL database
}
\examples{
correlationAnalyzeR::getCorrelationData(Species = "hsapiens",
                                    Sample_Type = "normal",
                                    Tissue = "blood",
                                    geneList = c("ATM", "BRCA1"))

}
