% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelationData.R
\name{getCorrelationData}
\alias{getCorrelationData}
\title{Get Gene Correlation Data}
\usage{
getCorrelationData(Sample_Type, Tissue, geneList, corrMat = NULL, pool = NULL)
}
\arguments{
\item{Sample_Type}{Type of RNA Seq samples used to create correlation data.
Either "all", "normal", or "cancer". Can be a single value for all genes,
or a vector corresponding to geneList. Not used if corrMat is set.}

\item{Tissue}{Which tissue type should gene correlations be derived from?
Default = "all". Can be a single value for all genes,
or a vector corresponding to geneList. Not used if corrMat is set.
Run getTissueTypes() to see available tissue list.}

\item{geneList}{Vector of genes for which data will be extracted.}

\item{corrMat}{A custom correlation matrix generated by generateCorrelations()
to use instead of pre-supplied databases.}

\item{pool}{an object created by pool::dbPool to accessing SQL database.
It will be created if not supplied. Not used if corrMat is set.}
}
\value{
A correlation data frame object
}
\description{
Obtain correlation data by querying MySQL database
}
\examples{
correlationAnalyzeR::getCorrelationData(Sample_Type = "normal",
                                       Tissue = "kidney",
                                       geneList = c("ATM", "BRCA1"))

}
