% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGSEA.R
\name{myGSEA}
\alias{myGSEA}
\title{Wrapper for clusterProlifer's GSEA function}
\usage{
myGSEA(ranks, TERM2GENE, plotFile, outDir, Condition = Condition,
  padjustedCutoff = 0.05)
}
\arguments{
\item{ranks}{Numeric of gene 'scores' ordered by decreasing value and
named with gene symbols.}

\item{TERM2GENE}{Data frame with two columns: gene set identifiers and
gene symbols.}

\item{plotFile}{prefix to use for naming output files.}

\item{outDir}{output directory.}

\item{Condition}{Name to use for titles of plots.}

\item{padjustedCutoff}{Value to use as a cutoff for returned gene sets.}
}
\value{
Output of GSEA function in clusterProfiler.
}
\description{
Run's the GSEA function from clusterProfiler and creates useful
    visualizations. It also outputs GSEA results to csv.
}
\examples{
myGSEA(ranks = DGE_results$GSEA, TERM2GENE = msigdbr()[,c(1,5)],
       plotFile = "GSEA_out", outDir = getwd(), Condition = "GSEA Results")

}
